fastapi_agent/pyproject.toml
[project]
name = "fastapi-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "Muhammad Fahad Khan", email = "fghory@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "agentops>=0.4.12",
    "chainlit>=2.5.5",
    "dotenv>=0.9.9",
    "fastapi[standard]>=0.115.12",
    "openai-agents[viz]>=0.0.16",
    "pydantic>=2.11.5",
    "tavily-python>=0.7.2",
]

[project.scripts]
panaversity = "fastapi_agent.panaversity:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
